Reorganize src folder structure by functionality
- Create organized directory structure:
  - Core/: audioSystem, AudioSequencer, audioDevice
  - Config/: ConfigReader, AudioConfig
  - Applications/: main.cpp, main_gui.cpp
  - Common/: notes.h, MidiEvent.h (shared headers)
  - Adapters/: AudioSystemAdapter
  - Effects/: (existing)
  - Envelope/: (existing)
  - Midi/: (existing)
  - Waves/: (existing)

- Update CMakeLists.txt files to reflect new paths
- Fix all #include statements to use new folder structure
- Remove duplicate MidiEvent.h from Midi folder
- Add utilities path to includes
- Both console and GUI apps build and run successfully

This improves code organization and maintainability by grouping
related functionality together.
