+++ src/CMakeLists.txt
@@ -1,10 +1,10 @@
 # Audio core library sources
 set(AUDIO_CORE_SOURCES
-    ConfigReader.cpp
-    audioSystem.cpp
-    audioDevice.cpp
-    AudioSequencer.cpp
-    AudioSystemAdapter.cpp
+    Config/ConfigReader.cpp
+    Core/audioSystem.cpp
+    Core/audioDevice.cpp
+    Core/AudioSequencer.cpp
+    Adapters/AudioSystemAdapter.cpp
     Midi/MidiDevice.cpp
     Effects/DelayEffect.cpp
     Effects/IEffect.cpp
@@ -22,7 +22,13 @@ add_library(audio_core OBJECT ${AUDIO_CORE_SOURCES})
 
 target_include_directories(audio_core PUBLIC
     ${CMAKE_CURRENT_SOURCE_DIR}
+    ${CMAKE_CURRENT_SOURCE_DIR}/Core
+    ${CMAKE_CURRENT_SOURCE_DIR}/Config
+    ${CMAKE_CURRENT_SOURCE_DIR}/Common
+    ${CMAKE_CURRENT_SOURCE_DIR}/Adapters
+    ${CMAKE_CURRENT_SOURCE_DIR}/Applications
     ${CMAKE_SOURCE_DIR}/src
+    ${CMAKE_SOURCE_DIR}/utilities
     ${RTAUDIO_INCLUDE_DIRS}
     ${RTMIDI_INCLUDE_DIRS}
     ${LIBXML2_INCLUDE_DIR}
